+++ djangojs/templatetags/js.py
@@ -175,10 +175,11 @@ def django_js(context, jquery=True, i18n=True, csrf=True):
 
 
 @register.inclusion_tag('djangojs/django_js_init.html', takes_context=True)
-def django_js_init(context, i18n=True, csrf=True):
+def django_js_init(context, jquery=False, i18n=True, csrf=True):
     '''Include Django.js javascript library initialization in the page'''
     return {
         'js': {
+            'jquery': _boolean(jquery),
             'i18n': _boolean(i18n),
             'csrf': _boolean(csrf),
         }